requirements.txt
python-dotenv==1.1.1
PyMuPDF==1.26.3
langchain==0.3.27
openai==1.98.0
google-generativeai==0.8.5
anthropic==0.25.0
pydantic==2.11.1
pyyaml==6.0.2
tiktoken==0.9.0
aiohttp==3.9.5
chardet==5.2.0
requests==2.32.3